requirements.txt
Django==4.2.7
firebase-admin==6.2.0
python-decouple==3.8
Pillow==10.0.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
whitenoise==6.6.0
gunicorn==21.2.0
